pyproject.toml
[tool.poetry]
name = "lucif"
version = "0.0.1"
description = "barb is gay"
authors = ["Pedro Antonio <anomalityinbox@gmail.com>"]
license = "GNU GPL V3"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
cogwatch = "^2.1.0"
python-dotenv = "^0.18.0"
jishaku = "^2.1.0"

[tool.poetry.scripts]
loader="lucif.loader:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
